+++ pytsite/assetman/__init__.py
@@ -3,8 +3,8 @@
 # Public API
 from . import _error as error
 from ._api import register_package, library, preload, remove, dump_js, dump_css, url, add_inline, dump_inline, \
-    get_urls, get_locations, reset, detect_collection, build, is_package_registered, register_global, _add_task, \
-    t_browserify, t_copy, t_copy_static, t_less, t_js, t_css, js_module, setup
+    get_urls, get_locations, reset, detect_collection, build, is_package_registered, register_global, t_browserify, \
+    t_copy, t_copy_static, t_less, t_js, t_css, js_module, setup, get_src_dir_path, get_dst_dir_path
 
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
@@ -21,6 +21,12 @@ def _init():
 
     # Resources
     lang.register_package(__name__)
+
+    # Do this here to avoid cyclic dependency
+    register_package('pytsite.lang')
+    t_js('pytsite.lang@**')
+    js_module('pytsite-lang-translations', 'pytsite.lang@translations')
+    js_module('pytsite-lang', 'pytsite.lang@pytsite-lang')
 
     # Console commands
     console.register_command(_console_command.Setup())